Updated Notification popup
I separated the notification function from the main.dart file and created a new "nofication.dart"file under the utilities folder.
